# server/openapi.yaml
openapi: 3.0.3
info:
  title: Campus Study Buddy API
  description: API for Campus Study Buddy - A platform for students to find study partners and form study groups
  version: 1.0.0
  contact:
    name: Campus Study Buddy Team
    email: support@campusstudybuddy.com
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.campusstudybuddy.com
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token
  schemas:
...
    Partner:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        modules:
          type: array
          items:
            type: string
        studyPreferences:
          type: array
          items:
            type: string
        availability:
          type: array
          items:
            type: string
        rating:
          type: number
          minimum: 0
          maximum: 5
        totalStudySessions:
          type: integer
          minimum: 0
...
    ProgressEntry:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        module:
          type: string
        topic:
          type: string
        hoursStudied:
          type: number
          minimum: 0
        confidence:
          type: integer
          minimum: 1
          maximum: 5
        notes:
          type: string
        date:
          type: string
          format: date-time
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    SearchSummary:
      type: object
      properties:
        filters:
          type: object
          properties:
            module:
              type: string
              nullable: true
            notModule:
              type: string
              nullable: true
            modules:
              type: string
              nullable: true
            modulesAll:
              type: string
              nullable: true
            page:
              type: integer
            limit:
              type: integer
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
        timestamp:
          type: string
          format: date-time
paths:
...
  /api/partners:
    get:
      summary: Get study partners with advanced filtering
      description: Retrieve a list of study partners with advanced filtering options including anti-search
      tags:
        - Partners
      security:
        - BearerAuth: []
      parameters:
        - name: module
          in: query
          description: Filter partners by single module code
          required: false
          schema:
            type: string
            example: COMS3011
        - name: notModule
          in: query
          description: Anti-search - exclude partners taking this module
          required: false
          schema:
            type: string
            example: COMS3011
        - name: modules
          in: query
          description: Filter by multiple modules (any match)
          required: false
          schema:
            type: string
            example: COMS3011,MATH2001
        - name: modulesAll
          in: query
          description: Filter by multiple modules (must match all)
          required: false
          schema:
            type: string
            example: COMS3011,MATH2001
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Maximum number of partners to return
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: List of study partners
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Partner'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'
                  searchSummary:
                    $ref: '#/components/schemas/SearchSummary'
                  message:
                    type: string
        '400':
          description: Bad request - Validation error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  details:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                        message:
                          type: string
                        value:
                          type: string
        '401':
          description: Unauthorized - Invalid or missing token
        '500':
          description: Internal server error
  /api/partners/{id}:
    get:
      summary: Get a specific study partner
      description: Retrieve details of a specific study partner by ID
      tags:
        - Partners
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Partner ID
          schema:
            type: string
      responses:
        '200':
          description: Partner details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Partner'
        '401':
          description: Unauthorized - Invalid or missing token
        '404':
          description: Partner not found
        '500':
          description: Internal server error
...
  /api/progress:
    get:
      summary: Get current user's progress summary
      description: Retrieve a summary of the current user's study progress
      tags:
        - Progress
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Progress summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      summary:
                        type: object
                        properties:
                          totalEntries:
                            type: integer
                          totalHours:
                            type: number
                          averageConfidence:
                            type: number
                          modules:
                            type: array
                            items:
                              type: string
                          lastUpdated:
                            type: string
                            format: date-time
                      recentEntries:
                        type: array
                        items:
                          $ref: '#/components/schemas/ProgressEntry'
        '401':
          description: Unauthorized - Invalid or missing token
        '500':
          description: Internal server error

    post:
      summary: Create a progress entry
      description: Create a new study progress entry
      tags:
        - Progress
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - module
                - topic
                - hours
              properties:
                module:
                  type: string
                  description: Module code
                  minLength: 1
                  maxLength: 20
                topic:
                  type: string
                  description: Study topic
                  minLength: 1
                  maxLength: 200
                hours:
                  type: number
                  minimum: 0.1
                  maximum: 24
                  description: Hours spent studying
                confidence:
                  type: integer
                  minimum: 1
                  maximum: 5
                  description: Confidence level (1-5)
                notes:
                  type: string
                  description: Additional notes
                  maxLength: 1000
      responses:
        '201':
          description: Progress entry created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/ProgressEntry'
                  message:
                    type: string
        '400':
          description: Bad request - Validation error
        '401':
          description: Unauthorized - Invalid or missing token
        '500':
          description: Internal server error
  /api/progress/{userId}:
    get:
      summary: Get user's progress
      description: Retrieve study progress for a specific user (users can only view their own progress)
      tags:
        - Progress
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: User ID
          schema:
            type: string
        - name: module
          in: query
          description: Filter by module code
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of entries to return
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: User's progress entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProgressEntry'
                  summary:
                    type: object
                    properties:
                      totalEntries:
                        type: integer
                      totalHours:
                        type: number
                      averageConfidence:
                        type: number
                      modules:
                        type: array
                        items:
                          type: string
                      lastUpdated:
                        type: string
                        format: date-time
        '400':
          description: Bad request - Validation error
        '401':
          description: Unauthorized - Invalid or missing token
        '403':
          description: Forbidden - Can only view own progress
        '500':
          description: Internal server error
...